requirements.txt
nltk~=3.5
sklearn
boto3==1.9.18
Django==2.1.3
djangorestframework==3.9.0
django-cors-headers==2.5.0
dlib==19.17.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
matplotlib==3.0.2
numpy==1.18
pandas==0.23.4
Pillow==5.0.0
requests==2.20.1
scipy==1.1.0
tensorflow
urllib3==1.24.1
gunicorn
imutils
scikit-learn==0.21.3
psutil
setproctitle
pyyaml~=5.3.1
bert-extractive-summarizer
spacy==2.1.3
transformers==2.2.2
neuralcoref
rake-nltk
torch
